Refactor : 서버사이드 selectQuizHandler 리팩토링
상세:
- selectQuizHandler의 로직을 모두 gameController로 이동
